+++ network/network_server/src/network_server.cpp
@@ -108,8 +108,8 @@ map<string, string> * ConnectionNetwork::RecvMsg()
 
 int ConnectionNetwork::SendFile(InFile * file_obj)
 {
-	int file_size = file_obj->GetSize();
-	int send_bytes = 0;
+	size_t file_size = file_obj->GetSize();
+	size_t send_bytes = 0;
 	std::array<char, chunkSize> buf;
 	buf.fill('\0');
 
@@ -125,8 +125,8 @@ int ConnectionNetwork::SendFile(InFile * file_obj)
 
 int ConnectionNetwork::RecvFile(OutFile * file_obj)
 {
-	int file_size = file_obj->GetSize();
-	int recv_bytes = 0;
+	size_t file_size = file_obj->GetSize();
+	size_t recv_bytes = 0;
 	std::array<char, chunkSize> buf;
 	buf.fill('\0');
